+++ include/SQLiteCpp/Statement.h
@@ -610,7 +610,7 @@ public:
     }
 
     // Return a UTF-8 string containing the SQL text of prepared statement with bound parameters expanded.
-    std::string getExtendedSQL();
+    std::string getExpandedSQL();
 
     /// Return the number of columns in the result set returned by the prepared statement
     inline int getColumnCount() const