+++ CHANGELOG.md
@@ -136,3 +136,4 @@ Version ?
 - #192 Add wrapper for bind parameter count
 - #197 Add tuple_bind and execute_many
 - #199 Fix #156 misleading error message in exception from Statement::exec
+- #201 Add Statement::getExpandedSQL() to get the SQL text of prepared statement with bound parameters expanded